requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
comm==0.2.2
crashtest==0.4.1
cryptography==44.0.2
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.10.0
fastapi==0.115.12
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.22.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
ipykernel==6.29.5
ipython==8.18.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
javalang==0.13.0
jedi==0.19.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
langchain==0.3.22
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.49
langchain-groq==0.3.2
langchain-ollama==0.3.0
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langgraph==0.3.22
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langsmith==0.3.22
marshmallow==3.26.1
matplotlib-inline==0.1.7
more-itertools==10.6.0
msgpack==1.1.0
multidict==6.3.1
mypy-extensions==1.0.0
nest_asyncio==1.6.0
numpy==2.0.2
ollama==0.4.7
openai==1.70.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
parso==0.8.4
pbs-installer==2025.3.17
pexpect==4.9.0
pickleshare==0.7.5
pip==25.0
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==5.8.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydantic-settings==2.8.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
stack_data==0.6.3
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
trove-classifiers==2025.3.19.19
typing_extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
wcwidth==0.2.13
wheel==0.45.1
wrapt==1.17.2
xattr==1.1.4
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
chardet~=5.2.0
tree-sitter~=0.23.2
tree-sitter-java~=0.23.5
aiofiles~=24.1.06
pydriller~=2.7